storage/fs: move SafeWriteToFile from storage
Move the SafeWriteToFile facility to the storage/fs subpackage. It's
directly related to the filesystem. The storage package's surface
area is very broad right now, and I'd like to trim it down by factoring
some items into subpackages.

Release note: None
