+++ pkg/storage/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "engine.go",
         "engine_key.go",
         "error.go",
-        "file_util.go",
         "in_mem.go",
         "intent_interleaving_iter.go",
         "intent_reader_writer.go",
@@ -97,7 +96,6 @@ go_test(
         "disk_map_test.go",
         "engine_key_test.go",
         "engine_test.go",
-        "file_util_test.go",
         "intent_interleaving_iter_test.go",
         "intent_reader_writer_test.go",
         "main_test.go",
